+++ Sources/Classes/core/MiniAppAnalytics.swift
@@ -12,6 +12,8 @@ internal enum MiniAppRATEvent: String, CaseIterable {
     case close
     case signatureFailure = "mini_app_signature_validation_fail"
     case getUniqueId = "mini_app_get_unique_id"
+    case getMessagingUniqueId = "mini_app_get_messaging_unique_id"
+    case getMauid = "mini_app_get_mauid"
     case getCurrentPosition = "mini_app_get_current_position"
     case requrestPermission = "mini_app_request_permission"
     case requrestCustomPermission = "mini_app_request_custom_permission"
@@ -44,6 +46,8 @@ internal enum MiniAppRATEvent: String, CaseIterable {
         case .open, .close:
             return .click
         case .getUniqueId,
+             .getMessagingUniqueId,
+             .getMauid,
              .getCurrentPosition,
              .requrestPermission,
              .requrestCustomPermission,
@@ -135,10 +139,15 @@ public class MiniAppAnalytics {
         sendAnalytics(event: ratEvent)
     }
 
+    // swiftlint:disable function_body_length
     private class func getRatEvent(for command: MiniAppJSActionCommand) -> MiniAppRATEvent? {
         switch command {
         case .getUniqueId:
             return .getUniqueId
+        case .getMessagingUniqueId:
+            return .getMessagingUniqueId
+        case .getMauid:
+            return .getMauid
         case .getCurrentPosition:
             return .getCurrentPosition
         case .requestPermission: